Cleanup and bugfixes

+++ core/src/main/java/com/nisovin/magicspells/Spellbook.java
@@ -41,7 +41,7 @@ public class Spellbook {
 		public boolean remove(Object o) {
 			boolean ret = super.remove(o);
 			if (o instanceof Spell) {
-				Spell s = (Spell)o;
+				Spell s = (Spell) o;
 				s.unloadPlayerEffectTracker(player);
 			}
 			return ret;
@@ -56,7 +56,7 @@ public class Spellbook {
 
 		@Override
 		public void clear() {
-			for (Spell s: this) {
+			for (Spell s : this) {
 				s.unloadPlayerEffectTracker(player);
 			}
 			super.clear();
@@ -64,7 +64,7 @@ public class Spellbook {
 
 	};
 
-	private Map<CastItem, ArrayList<Spell>> itemSpells = new HashMap<>();
+	private Map<CastItem, List<Spell>> itemSpells = new HashMap<>();
 	private Map<CastItem, Integer> activeSpells = new HashMap<>();
 	private Map<Spell, Set<CastItem>> customBindings = new HashMap<>();
 	private Map<Plugin, Set<Spell>> temporarySpells = new HashMap<>();
@@ -74,8 +74,8 @@ public class Spellbook {
 	public Spellbook(Player player, MagicSpells plugin) {
 		this.plugin = plugin;
 		this.player = player;
-		this.playerName = player.getName();
-		this.uniqueId = Util.getUniqueId(player);
+		playerName = player.getName();
+		uniqueId = Util.getUniqueId(player);
 
 		MagicSpells.debug(1, "Loading player spell list: " + playerName);
 		load();
@@ -104,7 +104,7 @@ public class Spellbook {
 		// Give all spells to ops, or if ignoring grant perms
 		if ((plugin.ignoreGrantPerms && plugin.ignoreGrantPermsFakeValue) || (player.isOp() && plugin.opsHaveAllSpells)) {
 			MagicSpells.debug(2, "  Op, granting all spells...");
-			for (Spell spell : plugin.spellsOrdered) {
+			for (Spell spell : MagicSpells.getSpellsOrdered()) {
 				if (spell.isHelperSpell()) continue;
 				if (!allSpells.contains(spell)) {
 					addSpell(spell);
@@ -117,12 +117,9 @@ public class Spellbook {
 
 		// Sort spells or pre-select if just one
 		for (CastItem i : itemSpells.keySet()) {
-			ArrayList<Spell> spells = itemSpells.get(i);
-			if (spells.size() == 1 && !plugin.allowCycleToNoSpell) {
-				activeSpells.put(i, 0);
-			} else {
-				Collections.sort(spells);
-			}
+			List<Spell> spells = itemSpells.get(i);
+			if (spells.size() == 1 && !plugin.allowCycleToNoSpell) activeSpells.put(i, 0);
+			else Collections.sort(spells);
 		}
 	}
 
@@ -133,53 +130,45 @@ public class Spellbook {
 			File file;
 			if (plugin.separatePlayerSpellsPerWorld) {
 				File folder = new File(plugin.getDataFolder(), "spellbooks" + File.separator + player.getWorld().getName());
-				if (!folder.exists()) {
-					folder.mkdir();
-				}
+				if (!folder.exists()) folder.mkdir();
+
 				file = new File(plugin.getDataFolder(), "spellbooks" + File.separator + playerWorld.getName() + File.separator + uniqueId + ".txt");
 				if (!file.exists()) {
 					File file2 = new File(plugin.getDataFolder(), "spellbooks" + File.separator + playerWorld.getName() + File.separator + playerName.toLowerCase() + ".txt");
-					if (file2.exists()) {
-						file2.renameTo(file);
-					}
+					if (file2.exists()) file2.renameTo(file);
 				}
 			} else {
 				file = new File(plugin.getDataFolder(), "spellbooks" + File.separator + uniqueId + ".txt");
 				if (!file.exists()) {
 					File file2 = new File(plugin.getDataFolder(), "spellbooks" + File.separator + playerName.toLowerCase() + ".txt");
-					if (file2.exists()) {
-						file2.renameTo(file);
-					}
+					if (file2.exists()) file2.renameTo(file);
 				}
 			}
 			if (file.exists()) {
 				Scanner scanner = new Scanner(file);
 				while (scanner.hasNext()) {
 					String line = scanner.nextLine();
-					if (!line.isEmpty()) {
-						if (!line.contains(":")) {
-							Spell spell = MagicSpells.getSpellByInternalName(line);
-							if (spell != null) {
-								addSpell(spell);
-							}
-						} else {
-							String[] data = line.split(":", 2);
-							Spell spell = MagicSpells.getSpellByInternalName(data[0]);
-							if (spell != null) {
-								ArrayList<CastItem> items = new ArrayList<>();
-								String[] s = data[1].split(",");
-								for (int i = 0; i < s.length; i++) {
-									try {
-										CastItem castItem = new CastItem(s[i]);
-										items.add(castItem);
-									} catch (Exception e) {
-										e.printStackTrace();
-									}
-								}
-								addSpell(spell, items.toArray(new CastItem[items.size()]));
-							}
+					if (line.isEmpty()) continue;
+					if (!line.contains(":")) {
+						Spell spell = MagicSpells.getSpellByInternalName(line);
+						if (spell != null) addSpell(spell);
+						continue;
+					}
+					String[] data = line.split(":", 2);
+					Spell spell = MagicSpells.getSpellByInternalName(data[0]);
+					if (spell == null) continue;
+
+					List<CastItem> items = new ArrayList<>();
+					String[] s = data[1].split(",");
+					for (int i = 0; i < s.length; i++) {
+						try {
+							CastItem castItem = new CastItem(s[i]);
+							items.add(castItem);
+						} catch (Exception e) {
+							e.printStackTrace();
 						}
 					}
+					addSpell(spell, items.toArray(new CastItem[items.size()]));
 				}
 				scanner.close();
 			}
@@ -193,7 +182,7 @@ public class Spellbook {
 	public void addGrantedSpells() {
 		MagicSpells.debug(2, "  Adding granted spells...");
 		boolean added = false;
-		for (Spell spell : plugin.spellsOrdered) {
+		for (Spell spell : MagicSpells.getSpellsOrdered()) {
 			MagicSpells.debug(3, "    Checking spell " + spell.getInternalName() + "...");
 			if (spell.isHelperSpell()) continue;
 			if (hasSpell(spell, false)) continue;
@@ -268,7 +257,7 @@ public class Spellbook {
 	}
 
 	public Set<Spell> getSpells() {
-		return this.allSpells;
+		return allSpells;
 	}
 
 	public List<String> tabComplete(String partial) {
@@ -282,14 +271,14 @@ public class Spellbook {
 				if (spell.isHelperSpell()) continue;
 				if (spell.getName().toLowerCase().startsWith(partial)) {
 					options.add(spell.getName());
-				} else {
-					String[] aliases = spell.getAliases();
-					if (aliases != null && aliases.length > 0) {
-						for (String alias : aliases) {
-							if (!alias.toLowerCase().startsWith(partial)) continue;
-							options.add(alias);
-						}
-					}
+					continue;
+				}
+
+				String[] aliases = spell.getAliases();
+				if (aliases == null || aliases.length <= 0) continue;
+				for (String alias : aliases) {
+					if (!alias.toLowerCase().startsWith(partial)) continue;
+					options.add(alias);
 				}
 			}
 			if (!options.isEmpty()) return options;
@@ -305,12 +294,9 @@ public class Spellbook {
 
 	protected CastItem getCastItemForCycling(ItemStack item) {
 		CastItem castItem;
-		if (item != null) {
-			castItem = new CastItem(item);
-		} else {
-			castItem = new CastItem(Material.AIR);
-		}
-		ArrayList<Spell> spells = itemSpells.get(castItem);
+		if (item != null) castItem = new CastItem(item);
+		else castItem = new CastItem(Material.AIR);
+		List<Spell> spells = itemSpells.get(castItem);
 		if (spells != null && (spells.size() > 1 || (spells.size() == 1 && plugin.allowCycleToNoSpell))) return castItem;
 		return null;
 	}
@@ -324,7 +310,7 @@ public class Spellbook {
 	protected Spell nextSpell(CastItem castItem) {
 		Integer i = activeSpells.get(castItem); // Get the index of the active spell for the cast item
 		if (i == null) return null;
-		ArrayList<Spell> spells = itemSpells.get(castItem); // Get all the spells for the cast item
+		List<Spell> spells = itemSpells.get(castItem); // Get all the spells for the cast item
 		if (!(spells.size() > 1 || i.equals(-1) || plugin.allowCycleToNoSpell || plugin.alwaysShowMessageOnCycle)) return null;
 		int count = 0;
 		while (count++ < spells.size()) {
@@ -357,7 +343,7 @@ public class Spellbook {
 	protected Spell prevSpell(CastItem castItem) {
 		Integer i = activeSpells.get(castItem); // Get the index of the active spell for the cast item
 		if (i == null) return null;
-		ArrayList<Spell> spells = itemSpells.get(castItem); // Get all the spells for the cast item
+		List<Spell> spells = itemSpells.get(castItem); // Get all the spells for the cast item
 		if (spells.size() > 1 || i.equals(-1) || plugin.allowCycleToNoSpell) {
 			int count = 0;
 			while (count++ < spells.size()) {
@@ -409,8 +395,7 @@ public class Spellbook {
 			save();
 			return true;
 		}
-		if (MagicSpells.plugin.enableTempGrantPerms && Perm.TEMPGRANT.has(player, spell)) return true;
-		return false;
+		return MagicSpells.plugin.enableTempGrantPerms && Perm.TEMPGRANT.has(player, spell);
 	}
 
 	public void addSpell(Spell spell) {
@@ -432,37 +417,34 @@ public class Spellbook {
 			CastItem[] items = spell.getCastItems();
 			if (castItems != null && castItems.length > 0) {
 				items = castItems;
-				HashSet<CastItem> set = new HashSet<>();
+				Set<CastItem> set = new HashSet<>();
 				for (CastItem item : items) {
 					if (item != null) set.add(item);
 				}
 				customBindings.put(spell, set);
-			} else if (plugin.ignoreDefaultBindings) {
-				return; // no cast item provided and ignoring default, so just stop here
-			}
+			} else if (plugin.ignoreDefaultBindings) return;
+
 			for (CastItem i : items) {
 				MagicSpells.debug(3, "        Cast item: " + i + (castItems != null ? " (custom)" : " (default)"));
-				if (i != null) {
-					ArrayList<Spell> temp = itemSpells.get(i);
-					if (temp != null) {
-						temp.add(spell);
-					} else {
-						temp = new ArrayList<>();
-						temp.add(spell);
-						itemSpells.put(i, temp);
-						activeSpells.put(i, plugin.allowCycleToNoSpell ? -1 : 0);
-					}
+				if (i == null) continue;
+				List<Spell> temp = itemSpells.get(i);
+				if (temp != null) {
+					temp.add(spell);
+					continue;
 				}
+				temp = new ArrayList<>();
+				temp.add(spell);
+				itemSpells.put(i, temp);
+				activeSpells.put(i, plugin.allowCycleToNoSpell ? -1 : 0);
 			}
 		}
 		// Remove any spells that this spell replaces
 		if (spell.replaces != null) {
 			for (String spellName : spell.replaces) {
 				Spell sp = MagicSpells.getSpellByInternalName(spellName);
-				if (sp != null) {
-					MagicSpells.debug(3, "        Removing replaced spell: " + sp.getInternalName());
-					removeSpell(sp);
-				}
+				if (sp == null) continue;
+				MagicSpells.debug(3, "        Removing replaced spell: " + sp.getInternalName());
+				removeSpell(sp);
 			}
 		}
 		// Prevent learning of spells this spell precludes
@@ -474,26 +456,20 @@ public class Spellbook {
 	}
 
 	public void removeSpell(Spell spell) {
-		if (spell instanceof BuffSpell) {
-			((BuffSpell)spell).turnOff(player);
-		}
+		if (spell instanceof BuffSpell) ((BuffSpell) spell).turnOff(player);
 		CastItem[] items = spell.getCastItems();
-		if (customBindings.containsKey(spell)) {
-			items = customBindings.remove(spell).toArray(new CastItem[]{});
-		}
+		if (customBindings.containsKey(spell)) items = customBindings.remove(spell).toArray(new CastItem[]{});
 		for (CastItem item : items) {
-			if (item != null) {
-				ArrayList<Spell> temp = itemSpells.get(item);
-				if (temp != null) {
-					temp.remove(spell);
-					if (temp.isEmpty()) {
-						itemSpells.remove(item);
-						activeSpells.remove(item);
-					} else {
-						activeSpells.put(item, -1);
-					}
-				}
+			if (item == null) continue;
+			List<Spell> temp = itemSpells.get(item);
+			if (temp == null) continue;
+			temp.remove(spell);
+			if (temp.isEmpty()) {
+				itemSpells.remove(item);
+				activeSpells.remove(item);
+				continue;
 			}
+			activeSpells.put(item, -1);
 		}
 		allSpells.remove(spell);
 	}
@@ -509,18 +485,15 @@ public class Spellbook {
 
 	public void removeTemporarySpells(Plugin plugin) {
 		Set<Spell> temps = temporarySpells.remove(plugin);
-		if (temps != null) {
-			for (Spell spell : temps) {
-				removeSpell(spell);
-			}
+		if (temps == null) return;
+		for (Spell spell : temps) {
+			removeSpell(spell);
 		}
 	}
 
 	private boolean isTemporary(Spell spell) {
 		for (Set<Spell> temps : temporarySpells.values()) {
-			if (temps.contains(spell)) {
-				return true;
-			}
+			if (temps.contains(spell)) return true;
 		}
 		return false;
 	}
@@ -532,7 +505,7 @@ public class Spellbook {
 		if (!bindings.contains(castItem)) bindings.add(castItem);
 
 		// Add to item bindings
-		ArrayList<Spell> bindList = itemSpells.get(castItem);
+		List<Spell> bindList = itemSpells.get(castItem);
 		if (bindList == null) {
 			bindList = new ArrayList<>();
 			itemSpells.put(castItem, bindList);
@@ -552,15 +525,13 @@ public class Spellbook {
 		}
 
 		// Remove from active bindings
-		ArrayList<Spell> bindList = itemSpells.get(castItem);
+		List<Spell> bindList = itemSpells.get(castItem);
 		if (bindList != null) {
 			removed = bindList.remove(spell) || removed;
 			if (bindList.isEmpty()) {
 				itemSpells.remove(castItem);
 				activeSpells.remove(castItem);
-			} else {
-				activeSpells.put(castItem, -1);
-			}
+			} else activeSpells.put(castItem, -1);
 		}
 
 		return removed;
@@ -574,7 +545,7 @@ public class Spellbook {
 
 	public void removeAllSpells() {
 		for (Spell spell : allSpells) {
-			if (spell instanceof BuffSpell) ((BuffSpell)spell).turnOff(player);
+			if (spell instanceof BuffSpell) ((BuffSpell) spell).turnOff(player);
 		}
 		allSpells.clear();
 		itemSpells.clear();
@@ -606,18 +577,17 @@ public class Spellbook {
 			}
 			BufferedWriter writer = new BufferedWriter(new FileWriter(file, false));
 			for (Spell spell : allSpells) {
-				if (!isTemporary(spell)) {
-					writer.append(spell.getInternalName());
-					if (customBindings.containsKey(spell)) {
-						Set<CastItem> items = customBindings.get(spell);
-						StringBuilder s = new StringBuilder();
-						for (CastItem i : items) {
-							s.append((s.length() == 0) ? "" : ",").append(i);
-						}
-						writer.append(String.valueOf(':')).append(s.toString());
+				if (isTemporary(spell)) continue;
+				writer.append(spell.getInternalName());
+				if (customBindings.containsKey(spell)) {
+					Set<CastItem> items = customBindings.get(spell);
+					StringBuilder s = new StringBuilder();
+					for (CastItem i : items) {
+						s.append((s.length() == 0) ? "" : ",").append(i);
 					}
-					writer.newLine();
+					writer.append(String.valueOf(':')).append(s.toString());
 				}
+				writer.newLine();
 			}
 			writer.close();
 			MagicSpells.debug(2, "Saved spellbook file: " + playerName.toLowerCase());
